Fix reporting of violations
There are several fixes here.

First of all, the `js_evaluation_allowed`
check now returns violations. Before, it would traverse all
directives and the `result` would always be the value of
the last directive checked. When you have both a regular CSP
header and the report-only header, it would only take into
account the last report-only header, effectively disabling
the enforcement.

Next to that, it now also fully implements the spec by creating
the violation following the spec by trimming the sample if
the report-sample directive is present.

Then, I discovered that the serialization of directives includes
a trailing whitespace. Some tests were asserting on the exact
value of the directive and thus find the extra space.

Lastly, integrity matches (which
`script-src-report-only-policy-works-with-external-hash-policy.html`
is using) were incorrectly implemented. They would skip for
`PolicyDisposition::Report`, but if you would have both an enforce
and report policy for this specific expression, it would incorrectly
skip. The integrity_matches algorithm is now implemented according
to the spec. I am not sure if the previous implementation was
accidentally correct, but I think it flipped the logic and it
coincidentally worked.